add exists check before removing submit code file in bash stdio evaluator
otherwise when there are multiple test cases, teardown for each test case will attempt to delete the file

+++ yaksh/bash_stdio_evaluator.py
@@ -25,7 +25,8 @@ class BashStdIOEvaluator(StdIOEvaluator):
         self.weight = test_case_data.get('weight')
 
     def teardown(self):
-        os.remove(self.submit_code_path)
+        if os.path.exists(self.submit_code_path):
+            os.remove(self.submit_code_path)
         if self.files:
             delete_files(self.files)
